# frostdb: fix BinaryScalarOperation equality operations