--- a/binaryscalarexpr_test.go
+++ b/binaryscalarexpr_test.go
@@ -5,6 +5,8 @@ import (
 
 	"github.com/segmentio/parquet-go"
 	"github.com/stretchr/testify/require"
+
+	"github.com/polarsignals/frostdb/query/logicalplan"
 )
 
 type FakeColumnChunk struct {
@@ -20,19 +22,18 @@ func (f *FakeColumnChunk) BloomFilter() parquet.BloomFilter { return nil }
 func (f *FakeColumnChunk) NumValues() int64                 { return 0 }
 
 type FakeColumnIndex struct {
-	numpages func() int
+	numPages int
+	min      parquet.Value
+	max      parquet.Value
 }
 
 func (f *FakeColumnIndex) NumPages() int {
-	if f.numpages != nil {
-		return f.numpages()
-	}
-	return 0
+	return f.numPages
 }
 func (f *FakeColumnIndex) NullCount(int) int64        { return 0 }
 func (f *FakeColumnIndex) NullPage(int) bool          { return false }
-func (f *FakeColumnIndex) MinValue(int) parquet.Value { return parquet.ValueOf(nil) }
-func (f *FakeColumnIndex) MaxValue(int) parquet.Value { return parquet.ValueOf(nil) }
+func (f *FakeColumnIndex) MinValue(int) parquet.Value { return f.min }
+func (f *FakeColumnIndex) MaxValue(int) parquet.Value { return f.max }
 func (f *FakeColumnIndex) IsAscending() bool          { return false }
 func (f *FakeColumnIndex) IsDescending() bool         { return false }
 
@@ -42,9 +43,7 @@ func (f *FakeColumnIndex) IsDescending() bool         { return false }
 func Test_MinMax_EmptyColumnChunk(t *testing.T) {
 	fakeChunk := &FakeColumnChunk{
 		index: &FakeColumnIndex{
-			numpages: func() int {
-				return 10
-			},
+			numPages: 10,
 		},
 	}
 
@@ -53,4 +52,69 @@ func Test_MinMax_EmptyColumnChunk(t *testing.T) {
 
 	v = Max(fakeChunk)
 	require.True(t, v.IsNull())
+}
+
+func TestBinaryScalarOperation(t *testing.T) {
+	for _, tc := range []struct {
+		name        string
+		min         int
+		max         int
+		right       int
+		op          logicalplan.Op
+		expectFalse bool
+	}{
+		{
+			name:  "OpEqValueContained",
+			min:   1,
+			max:   10,
+			right: 5,
+			op:    logicalplan.OpEq,
+		},
+		{
+			name:        "OpEqValueGt",
+			min:         1,
+			max:         10,
+			right:       11,
+			op:          logicalplan.OpEq,
+			expectFalse: true,
+		},
+		{
+			name:        "OpEqValueLt",
+			min:         1,
+			max:         10,
+			right:       0,
+			op:          logicalplan.OpEq,
+			expectFalse: true,
+		},
+		{
+			name:  "OpEqMaxBound",
+			min:   1,
+			max:   10,
+			right: 10,
+			op:    logicalplan.OpEq,
+		},
+		{
+			name:  "OpEqMinBound",
+			min:   1,
+			max:   10,
+			right: 1,
+			op:    logicalplan.OpEq,
+		},
+	} {
+		t.Run(tc.name, func(t *testing.T) {
+			if tc.op == logicalplan.OpUnknown {
+				t.Fatal("test programming error: remember to set operator")
+			}
+			fakeChunk := &FakeColumnChunk{
+				index: &FakeColumnIndex{
+					numPages: 1,
+					min:      parquet.ValueOf(tc.min),
+					max:      parquet.ValueOf(tc.max),
+				},
+			}
+			res, err := BinaryScalarOperation(fakeChunk, parquet.ValueOf(tc.right), tc.op)
+			require.NoError(t, err)
+			require.Equal(t, !tc.expectFalse, res)
+		})
+	}
 }