This commit fixes equality operations against ColumnChunks. The operation would previously return that a value in the column chunk might return true when in fact the column chunk could be discarded.

This greatly increases performance because we avoid decompressing unnecessary row groups:
```
name            old time/op    new time/op    delta
Query/Merge-8     61.0ms ± 2%    44.9ms ± 3%  -26.42%  (p=0.000 n=10+10)
Query/Range-8     57.9ms ± 1%    30.4ms ± 2%  -47.60%  (p=0.000 n=10+10)

name            old alloc/op   new alloc/op   delta
Query/Merge-8     1.06GB ± 0%    0.80GB ± 0%  -23.88%  (p=0.000 n=10+10)
Query/Range-8      190MB ± 0%      95MB ± 0%  -49.94%  (p=0.000 n=10+8)

name            old allocs/op  new allocs/op  delta
Query/Merge-8       917k ± 0%      445k ± 0%  -51.47%  (p=0.000 n=10+10)
Query/Range-8      4.82M ± 0%     2.34M ± 0%  -51.30%  (p=0.000 n=10+10)
```

This should also reduce object storage reads for persistent queries.